Clean up the datagrid CSS organization and add namespaced variables.
This updates the datagrids code to introduce a `#rb-ns-ui.datagrids`
namespace containing the various margins, paddings, colors, and other
styles used by our datagrid. This builds upon the new
`#djblets-ns-ui.datagrids` namespace in Djblets, extending those to give
us a single place where we can look up all variables.

Some of the datagrid code has been shuffled around in the process. The
datagrid classes have been made flatter and more reusable. Some styles
had to be updated to account for this, as specificity changed.

It's not a full clean-up, and documentation is not being added at this
stage. Eventually we'll want to rethink the component structure and
styling of datagrids in Djblets, and rethink some of the add-on classes
we've introduced here, but that's a change for another day.

Testing Done:
Checked all the datagrids and the affected columns, making sure it all
appeared to render correctly.

Reviewed at https://reviews.reviewboard.org/r/10665/