Set the correct root URL in RB.TextBasedReviewableView
The `RB.TextBasedReviewableView` was using an incorrect root path
Instead of just being the path relative to the FQDN, it was using the
entire URL. This resulted in Backbone never matching URLs since it was
trying to match fragments against full URLs. The view would then be
stuck in the rendered view mode at all times.

Testing Done:
Verified that the TextBasedReviewableView can switch between rendered
and source modes on Python 2.7, 3.5, 3.6, and 3.7.

Reviewed at https://reviews.reviewboard.org/r/10701/