Change the way that StarManagerView listens for clicks.
The StarManagerView used Backbone's `events` object to listen for clicks
on individual stars. Unfortunately the way that these events are bound
caused some event handler ordering issues with other work on the
dashboard to support clicking on cells which don't contain explicit
links. Changing it so that handler is attached using `$.on` directly on
the element fixes this.

Testing Done:
Toggled stars in the dashboard and on review request pages.

Reviewed at https://reviews.reviewboard.org/r/10711/