Add a CSS component for search fields.
This gives us a search field we can use in multiple locations. We've
previously been re-crafting search fields over and over, which doesn't
make much sense.

The field is basic. It consists of an outer container, an icon, and a
text input field.

As of this change, nothing currently uses this component, but an
upcoming change will move the New Review Request page over to it.

Testing Done:
Used this in the upcoming New Review Request page rewrite.

Reviewed at https://reviews.reviewboard.org/r/10741/