Allow subclasses of RB.CollectionView to specify the item container element.
`RB.CollectionView` makes automatic rendering of items from a collection
easy, but it has a limitation in that all items render to the
collection's main element. This prevents subclasses from having a more
complex element tree.

This updates `RB.CollectionView` to introduce a new `$container`
attribute, which is used for all rendering. It defauls to `$el`, but
subclasses or consumers can override it to point to another element in
the view.

Testing Done:
Unit tests passed.

Successfully used this new attribute in an upcoming change.

Reviewed at https://reviews.reviewboard.org/r/10740/