Update the datagrid pages to use the new setupPage JS support.
Review Board 4.0 introduced a new consistent way of setting up page
models and views in JavaScript, which coordinates the initialization and
events of page elements. This change moves all the datagrid pages over
to the new method.

To do this, the base datagrid template now fills in the page template
blocks to specify the model/view classes that the datagrid wants, along
with the attributes and options for those classes. This makes use of new
methods in `DataGridJSMixin`, which subclasses can override to provide
additional information.

Testing Done:
Tested that all the datagrid pages were still working as expected in mobile
and desktop modes.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10736/