Fix scrolling the sidebar in mobile mode.
On pages without full-page content mode, the page wasn't properly set up
to allow the mobile sidebar to scroll. The sidebar's height was bound to
the `#container` height, which would be based on the content height.
Only in full-page content mode would the `#container` height be bound to
the viewport.

This adds a couple of extra rules to activate the correct container
height when the mobile menu is open, and to always enable sidebar
scrolling in mobile mode, ensuring that the UI always behaves as
expected regardless of the modes used on the page.

Testing Done:
Tested the mobile scrollbar behavior on mobile mode in pages with both
standard and full-page content modes.

Reviewed at https://reviews.reviewboard.org/r/10750/