Add base styles and page layout for the new admin UI.
Review Board 4.0 is getting a fresh new administration UI. This is long
overdue, and is a requirement for modern Django support (as much has
changed in the Django admin UI styles and page layout since 1.6). The
goal is to produce a new admin UI that feels like the rest of Review
Board, is mobile-friendly, and is less reliant on Django's
implementation while also being compatible with Django apps.

This initial change introduces some base definitions, template blocks,
and JavaScript that future changes will build upon. It moves to the new
sidebar support, giving us both a mobile and desktop sidebar. It also
removes the page tabs in favor of entries in the sidebar, removing the
odd split in navigation that we've had for years.

The content areas in the admin UI have not really changed yet, and may
be in flux as this project comes together. Upcoming changes will iterate
on these pages until the entire admin UI is consistent and future-proof,
at which point we'll be able to drop the Django styling and finally
upgrade to Django 1.11.

Testing Done:
Tested that the sidebar entries all work.

Tested general layout in mobile and desktop, on Django 1.6 and 1.11.

Reviewed at https://reviews.reviewboard.org/r/10748/