Add and make use of a constant in settings for the product name.
Throughout the product, we've always hard-coded the product name as
"Review Board", which we then had to override in RBCommons, often
resulting in a mix of names. Sometimes this is appropriate (as we're not
hiding that the product is Review Board underneath), but often it is
not and can end up being a bit confusing.

This adds a new `settings.PRODUCT_NAME` constant that we can use
throughout the product. It's defined as "Review Board" here, but we'll
be able to override that with "RBCommons" and reduce the number of
templates and other code paths we have to override, making future merges
easier.

Testing Done:
Tested all affected pages with and without a custom product name.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10751/