Improve appearance, terminology, state, and options for page sidebars.
This adds another level of polish on top of the new page sidebars,
adding some breathing room and making the connected appearance optional.
The padding around the page sidebars has been lessened, giving more
space to the text content in each item, and creating a visual
consistency when the sidebar is not in a connected state.

The connected state, where the items appear to connect to a box on the
page, is now optional, and off by default. Pages have to opt into this.
This ensures that a page's sidebar will always look reasonable by
default. The datagrids and New Review Request page enable this state,
but the new admin UI's dashboard does not. The state is also always
enabled when in mobile mode, allowing the items to always appear to be
connected to the page content area.

The header for Review Board's main navigation items in the mobile
sidebar has also been renamed. It now uses the product name instead of
"Other Pages," which is hopefully more clear, and better matches what
you see in desktop mode.

And finally, there are fixes for the actual display and default items in
the page sidebar. The page sidebar elements were always visible on the
page, even if empty and without `-has-sidebar` being set. The items
populated weren't correct when anonymous, due to bad placement for
conditionals. Both are now fixed.

Testing Done:
Tested the sidebar in desktop and mobile modes on the dashboard,
user page, New Review Request page, review request page, and administration
UI.

Reviewed at https://reviews.reviewboard.org/r/10752/