Add API support for filtering review requests by branch.
This adds a `?branch=` filter to the review request API, allowing review
requests to be filtered by branch. This compares directly, and is
case-sensitive.

At the moment, this does not make use of any database indexes, so for
*very* large installs, this filter may be slow. We are not providing a
database index for branch fields at this time, in order to avoid
impacting the upgrade process for these servers, but if administrators
need to make use of this API and are seeing slowdown, an index on the
branch column can be added manually. We will be indexing this branch in
Review Board 4.0.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10763/