Add API for updating a user's name, e-mail, and is_active flag.
This introduces support for an HTTP PUT to the User resource for
updating a full name, e-mail address, and the `is_active` flag. The full
name and e-mail address can be updated only if the auth backend supports
it, and if it does then the backend will be notified of any changes.

Users can modify only their own full name and e-mail address.
Administrators and special users with the `auth.change_user` permission
can also update these fields for other users, along with the `is_active`
flag.

There's also a fix for e-mail validation errors when creating a new user.

At this time, user settings and `extra_data` cannot be modified or
retrieved via the API, as we need to consider whether users should be
able to see this state for other users and whether there could be any
security impacts.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10765/