Add a function for getting the Repository for a FileDiff.
This adds `FileDiff.get_repository()`, which returns the `Repository`
associated with a `FileDiff`. At the moment, this returns
`self.diffset.repository`, preserving the existing behavior completely.

In a future change, it will be possible for a `FileDiff` to be mapped to
a separate repository, paving the way for multi-repository diffs.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10776/